i18n added and error url changed
