--- designer/client/pages/ErrorPage/index.ts
@@ -1,1 +1,0 @@
-export { default as ErrorPage } from "./ErrorPage";
